cloud-backend/stt-worker/requirements.txt
deepgram-sdk==3.7.7
redis[asyncio]==5.1.1
asyncpg==0.29.0
python-dotenv==1.0.1
